requirements.txt
# 核心依赖
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0

# 数据库
clickhouse-driver>=0.2.0

# API 客户端
python-binance>=1.0.16

# Web 服务
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# 机器学习
scikit-learn>=1.0.0
torch>=1.10.0
optuna>=2.10.0
ta-lib>=0.4.0  # 技术分析库
ta>=0.9.0      # 备用技术分析库

# 工具
python-dotenv>=0.19.0
loguru>=0.5.3
plotly>=5.3.0
tqdm>=4.62.0
distro>=1.7.0  # 用于字体安装脚本检测Linux发行版

# 测试
pytest>=6.2.5

# 文档
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0

# 开发工具
black>=21.9b0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910 